// AngularClient/src/app/models/user-interface.ts
export class EmpleadoModel {
  constructor(){
    this.usuario = null;
    this.nombre = null;
    this.apellido = null;
    this.tipoidentificacion = null;
    this.numeroidentificacion = null;
    this.pais = null;
    this.correo = null;
    this.estado = null;
  }
  codu: number;
  usuario: string;
  nombre: string;
  apellido: string;
  tipoidentificacion: string;
  numeroidentificacion: string;
  pais: string;
  correo: string;
  estado: number;
}